// shared/addresses/types.ts
export interface AddressWordBase {
  nodeType: "word";
  value: string;
}

export interface AddressNodeWithUnclassifiedWord extends AddressWordBase {
  wordType: "unclassified";
}
// ...
export interface AddressNodeWithSeparator {
  nodeType: "separator";
  separatorType: "comma" | "slash" | "dash";
}

export type CleanedAddressNode = AddressNodeWithWord | AddressNodeWithSeparator;
export interface CleanedAddressAst {
  nodeType: "cleanedAddress";
  children: CleanedAddressNode[];
}

export interface WordReplacementConfig {
  detached?: boolean;
  from: string | string[];
  to: string;
  silenceNormalizationWarning?: boolean;
}

export interface WordReplacementDirective {
  detached: boolean;
  from: AddressNodeWithUnclassifiedWord[];
  to: AddressNodeWithUnclassifiedWord[];
}

export interface WordReplacementDirectiveTree {
  directive?: WordReplacementDirective;
  subtreeByWordValue: Record<string, WordReplacementDirectiveTree>;
}